+++ packages/synapse-interface/components/Banner.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 
-const BANNER_VERSION = '1'
+const BANNER_VERSION = '2'
 
 export const Banner = () => {
   const [hasMounted, setHasMounted] = useState(false)
@@ -47,7 +47,15 @@ export const Banner = () => {
         <div className="m-1 font-thin">
           <div className="container mx-auto">
             <p className="text-md">
-              New! Chain and token selections now list routable options first.
+              {/* TODO: replace w/ tweetlink directly */}
+              <a
+                href={
+                  'https://twitter.com/SynapseProtocol/status/1701315115692466220'
+                }
+                target={'_blank'}
+              >
+                Synapse Interchain Network & Synapse Chain are live on testnet.
+              </a>
             </p>
           </div>
         </div>
@@ -57,7 +65,7 @@ export const Banner = () => {
             inline-flex items-center justify-center
             h-7 w-7
             ml-auto -mx-1.5 -my-1.5 p-1.5
-            text-primaryTextColor 
+            text-primaryTextColor
           `}
           data-dismiss-target="#banner-default"
           aria-label="Close"