+++ packages/synapse-interface/components/Warning.tsx
@@ -1,109 +1,32 @@
+import { useMemo } from 'react'
+
 import { DISCORD_URL, TWITTER_URL } from '@/constants/urls'
-import {
-  ARBITRUM,
-  AVALANCHE,
-  DOGE,
-  ETH,
-  FANTOM,
-  HARMONY,
-} from '@/constants/chains/master'
+import { AVALANCHE, DOGE, FANTOM, HARMONY } from '@/constants/chains/master'
 import { useBridgeState } from '@/slices/bridge/hooks'
+import { CCTP_ROUTER_ADDRESS } from '@/utils/actions/fetchPortfolioBalances'
 
 export const Warning = () => {
-  const { fromChainId, toChainId, fromToken, toToken } = useBridgeState()
+  const { fromChainId, toChainId, fromToken, bridgeQuote } = useBridgeState()
 
-  const fromTokenSymbol = fromToken && fromToken?.symbol
-  const toTokenSymbol = toToken && toToken?.symbol
-
-  const isTokenUSDCAndUSDCe =
-    (fromToken?.symbol === 'USDC' && toToken?.symbol === 'USDCe') ||
-    (fromToken?.symbol === 'USDCe' && toToken?.symbol === 'USDC')
-
-  const isChainAvalancheOrArbitrum =
-    fromChainId === AVALANCHE.id ||
-    fromChainId === ARBITRUM.id ||
-    toChainId === AVALANCHE.id ||
-    toChainId === ARBITRUM.id
-
-  const isTokenUSDC = fromTokenSymbol === 'USDC' && toTokenSymbol === 'USDC'
-
-  const isChainOtherThanAvalancheArbitrumEthereum =
-    (toChainId != AVALANCHE.id &&
-      toChainId != ARBITRUM.id &&
-      toChainId != ETH.id) ||
-    (fromChainId != AVALANCHE.id &&
-      fromChainId != ARBITRUM.id &&
-      fromChainId != ETH.id)
-
-  const isTokenUSDCAndChainEthereumArbitrumAvalanche =
-    isTokenUSDC &&
-    ((fromChainId === ETH.id && toChainId === ARBITRUM.id) ||
-      (fromChainId === ARBITRUM.id && toChainId === AVALANCHE.id) ||
-      (fromChainId === AVALANCHE.id && toChainId === ARBITRUM.id))
+  const isCCTP = useMemo(() => {
+    return (
+      fromChainId &&
+      fromToken &&
+      fromChainId !== AVALANCHE.id &&
+      fromToken?.swapableType === 'USD' &&
+      bridgeQuote.routerAddress.toLowerCase() ===
+        CCTP_ROUTER_ADDRESS.toLowerCase()
+    )
+  }, [bridgeQuote, fromToken])
 
   const isChainHarmony = [fromChainId, toChainId].includes(HARMONY.id)
   const isChainFantom = [fromChainId, toChainId].includes(FANTOM.id)
   const isChainDoge = [fromChainId, toChainId].includes(DOGE.id)
 
-  if (isTokenUSDCAndUSDCe && isChainAvalancheOrArbitrum) {
+  if (isCCTP) {
     return (
       <WarningMessage
-        header="USDC and USDC.e are incompatible on some routes."
-        message={
-          <>
-            <p className="mb-2">
-              If you don't see an output, try bridging to another stablecoin.
-              For example, on some routes, USDC can currently only be bridged
-              with itself, via the new Circle cross-chain transfer protocol.
-            </p>
-            <p>
-              Follow{' '}
-              <a target="_blank" className="underline" href={TWITTER_URL}>
-                Twitter
-              </a>{' '}
-              or{' '}
-              <a target="_blank" className="underline" href={DISCORD_URL}>
-                Discord
-              </a>{' '}
-              for updates as more CCTP routes become available.
-            </p>
-          </>
-        }
-      />
-    )
-  } else if (
-    isTokenUSDC &&
-    isChainAvalancheOrArbitrum &&
-    isChainOtherThanAvalancheArbitrumEthereum
-  ) {
-    return (
-      <WarningMessage
-        header="Native USDC on Avalanche and Arbitrum may not be bridgeable to most chains, try another stablecoin instead."
-        message={
-          <>
-            <p className="mb-2">
-              Synapse is currently opening liquidity routes for Circle's Native
-              USDC to other chains.
-            </p>
-            <p>
-              Follow{' '}
-              <a target="_blank" className="underline" href={TWITTER_URL}>
-                Twitter
-              </a>{' '}
-              or{' '}
-              <a target="_blank" className="underline" href={DISCORD_URL}>
-                Discord
-              </a>{' '}
-              for updates when this route becomes available.
-            </p>
-          </>
-        }
-      />
-    )
-  } else if (isTokenUSDCAndChainEthereumArbitrumAvalanche) {
-    return (
-      <WarningMessage
-        header="USDC now uses the Circle cross-chain transfer protocol."
+        header="This route uses the Circle cross-chain transfer protocol."
         message={
           <>
             <p className="mb-2">