+++ packages/synapse-interface/constants/tokens/poolMaster.ts
@@ -20,6 +20,7 @@ import {
   ONEDAI,
   ONEUSDC,
   ONEUSDT,
+  WETH,
 } from '@constants/tokens/bridgeable'
 import {
   AVWETH,
@@ -29,7 +30,6 @@ import {
   KLAYTN_oUSDT,
   MIM,
   MULTIAVAX,
-  WETH,
 } from '@constants/tokens/auxilliary'
 import synapseLogo from '@assets/icons/syn.svg'
 
@@ -504,7 +504,7 @@ export const KLAYTN_ORBIT_SWAP_TOKEN = new Token({
   swapAddresses: {
     [CHAINS.KLAYTN.id]: '0xfDbaD1699A550F933EFebF652a735F2f89d3833c',
   },
-  poolTokens: [KLAYTN_oUSDT],
+  poolTokens: [USDT, KLAYTN_oUSDT],
   description: "Synapse's 2pool stableswap LP token on Klaytn",
   display: true,
   priorityPool: true,