+++ packages/synapse-interface/slices/portfolio/actions.ts
@@ -4,8 +4,8 @@ import { createAction } from '@reduxjs/toolkit'
 import { Token } from '@/utils/types'
 
 export enum PortfolioTabs {
-  HOME = 'home',
   PORTFOLIO = 'portfolio',
+  ACTIVITY = 'activity',
 }
 
 export enum FetchState {